将SDRAM控制器类的内存大小类型从int32_t改为int64_t以支持更大数据量

+++ include/base/embedded/sdram/chip/w9825g6kh_6/W9825G6KH_6.h
@@ -88,7 +88,7 @@ namespace base
 					{
 						return base::Span{
 							_controller.Span().Buffer(),
-							std::min<int32_t>(32 * 1024 * 1024, _controller.Span().Size()),
+							std::min<int64_t>(32 * 1024 * 1024, _controller.Span().Size()),
 						};
 					}
 				};
